chore(notification-dismiss-button): move to common/Notifications/
