+++ packages/components/src/Toast/Toast.tsx
@@ -1,4 +1,4 @@
-import NotificationDismissButton from "../common/NotificationDismissButton";
+import NotificationDismissButton from "../common/Notifications/NotificationDismissButton";
 import NotificationIcon from "../common/Notifications/NotificationIcon";
 import React from "react";
 import Text from "../Text";
